feat: Moved prompts to separate files
