+++ packages/create-liveblocks-app/bin/templates/example/index.ts
@@ -1,1 +1,1 @@
-export * from "./exampleTemplate";
+export * from "./example-template";
